Fixing merge conflict Working on Lauren's comment
changed assignee to map user DTO
removing comment

+++ packages/api/src/main/java/com/nsa/bt/scrumble/models/Workspace.java
@@ -11,7 +11,7 @@ public class Workspace {
     private ArrayList<Integer> projectIds;
     private List<User> users;
 
-    public Workspace(long id, User createdBy, String name, String description, ArrayList<Integer> projectIds, List<User> users) {
+    public Workspace(int id, User createdBy, String name, String description, ArrayList<Integer> projectIds, List<User> users) {
         this.id = id;
         this.createdBy = createdBy;
         this.name = name;
